Updated create_prayertime.py
Because the code relied on the website (muslimpro.com) format of date and times, the code broke when the website uses two digits for days 10-30/31 of the month and only a single digit for days 1-9. So I implemented a check, checking whether a single or double digits are used so that the times can be formatted correctly for the google api to understand and place on my calendar. #surprisinglythiswasfun